chore: change ownership of profile sync from notifications to identity
