+++ app/actions/identity/constants/errors.ts
@@ -1,0 +1,8 @@
+export enum identityErrors {
+  PERFORM_SIGN_IN = 'Error while trying to sign in',
+  PERFORM_SIGN_OUT = 'Error while trying to sign out',
+  ENABLE_PROFILE_SYNCING = 'Error while trying to enable profile syncing',
+  DISABLE_PROFILE_SYNCING = 'Error while trying to disable profile syncing',
+}
+
+export default identityErrors;
