+++ app/selectors/identity/index.tsx
@@ -1,0 +1,41 @@
+import { createSelector } from 'reselect';
+import { RootState } from '../../reducers';
+import {
+  AuthenticationController,
+  UserStorageController,
+} from '@metamask/profile-sync-controller';
+
+type AuthenticationState =
+  AuthenticationController.AuthenticationControllerState;
+type UserStorageState = UserStorageController.UserStorageControllerState;
+
+const selectAuthenticationControllerState = (state: RootState) =>
+  state?.engine?.backgroundState?.AuthenticationController ??
+  AuthenticationController.defaultState;
+
+const selectUserStorageControllerState = (state: RootState) =>
+  state?.engine?.backgroundState?.UserStorageController ??
+  UserStorageController.defaultState;
+
+export const selectIsProfileSyncingEnabled = createSelector(
+  selectUserStorageControllerState,
+  (userStorageControllerState: UserStorageState) =>
+    userStorageControllerState?.isProfileSyncingEnabled,
+);
+export const selectIsProfileSyncingUpdateLoading = createSelector(
+  selectUserStorageControllerState,
+  (userStorageControllerState: UserStorageState) =>
+    userStorageControllerState.isProfileSyncingUpdateLoading,
+);
+
+export const selectIsSignedIn = createSelector(
+  selectAuthenticationControllerState,
+  (authenticationControllerState: AuthenticationState) =>
+    authenticationControllerState.isSignedIn,
+);
+
+export const selectSessionData = createSelector(
+  selectAuthenticationControllerState,
+  (authenticationControllerState: AuthenticationState) =>
+    authenticationControllerState.sessionData,
+);
