+++ app/util/identity/hooks/useCreateSession.test.tsx
@@ -7,8 +7,8 @@ import initialRootState from '../../../util/test/initial-root-state';
 import React from 'react';
 import { Provider } from 'react-redux';
 import createMockStore from 'redux-mock-store';
-import * as Selectors from '../../../selectors/notifications';
-import * as Actions from '../../../actions/notification/helpers';
+import * as Selectors from '../../../selectors/identity';
+import * as Actions from '../../../actions/identity';
 import useCreateSession from './useCreateSession';
 
 jest.mock('../constants', () => ({
@@ -55,7 +55,9 @@ function arrangeSelectors() {
 }
 
 function arrangeActions() {
-  const mockSignIn = jest.spyOn(Actions, 'signIn').mockResolvedValue(undefined);
+  const mockSignIn = jest
+    .spyOn(Actions, 'performSignIn')
+    .mockResolvedValue(undefined);
   const mockDisableProfileSyncing = jest
     .spyOn(Actions, 'disableProfileSyncing')
     .mockResolvedValue(undefined);
