+++ app/util/identity/hooks/useCreateSession.ts
@@ -1,17 +1,21 @@
 import { useCallback, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { UseCreateSessionReturn } from './types';
 import { getErrorMessage } from '../../../util/errorHandling';
 
 import {
   selectIsProfileSyncingEnabled,
   selectIsSignedIn,
-} from '../../../selectors/notifications';
+} from '../../../selectors/identity';
 import {
   disableProfileSyncing,
-  signIn,
-} from '../../../actions/notification/helpers';
-import { isNotificationsFeatureEnabled } from '../constants';
+  performSignIn,
+} from '../../../actions/identity';
+
+interface UseCreateSessionReturn {
+  createSession: () => Promise<void>;
+  loading: boolean;
+  error?: string;
+}
 
 /**
  * Custom hook to manage the creation of a session based on the user's authentication status,
@@ -32,10 +36,6 @@ function useCreateSession(): UseCreateSessionReturn {
   const [error, setError] = useState<string | undefined>(undefined);
 
   const createSession = useCallback(async () => {
-    if (!isNotificationsFeatureEnabled()) {
-      return;
-    }
-
     // If the user is already signed in, no need to create a new session
     if (isSignedIn) {
       return;
@@ -51,7 +51,7 @@ function useCreateSession(): UseCreateSessionReturn {
       setLoading(true);
       setError(undefined);
       try {
-        const errorMessage = await signIn();
+        const errorMessage = await performSignIn();
         if (errorMessage) {
           throw new Error(errorMessage);
         }
