+++ app/util/identity/hooks/useProfileSyncing.test.tsx
@@ -7,7 +7,7 @@ import { act, renderHook } from '@testing-library/react-hooks';
 import React from 'react';
 import { Provider } from 'react-redux';
 import createMockStore from 'redux-mock-store';
-import * as Actions from '../../../actions/notification/helpers';
+import * as Actions from '../../../actions/identity';
 import initialRootState from '../../../util/test/initial-root-state';
 import { useProfileSyncing } from './useProfileSyncing';
