+++ app/util/identity/hooks/useProfileSyncing.ts
@@ -1,11 +1,17 @@
 import { useState, useCallback } from 'react';
-import { ProfileSyncingReturn } from './types';
 import { getErrorMessage } from '../../../util/errorHandling';
 import {
   disableProfileSyncing as disableProfileSyncingAction,
   enableProfileSyncing as enableProfileSyncingAction,
-} from '../../../actions/notification/helpers';
-import { isNotificationsFeatureEnabled } from '../constants';
+} from '../../../actions/identity';
+import { isNotificationsFeatureEnabled } from '../../notifications/constants';
+
+interface UseProfilSyncingReturn {
+  enableProfileSyncing: () => Promise<string | undefined>;
+  disableProfileSyncing: () => Promise<string | undefined>;
+  loading: boolean;
+  error?: string;
+}
 
 /**
  * Custom hook to enable profile syncing. This hook handles the process of signing in
@@ -13,7 +19,7 @@ import { isNotificationsFeatureEnabled } from '../constants';
  *
  * @returns An object containing the `enableProfileSyncing` function, loading state, and error state.
  */
-export function useProfileSyncing(): ProfileSyncingReturn {
+export function useProfileSyncing(): UseProfilSyncingReturn {
   const [loading, setLoading] = useState<boolean>(false);
   const [error, setError] = useState<string>();
